Move experimental 'Konfig' class to separate module
